Pass Brightness value in range 0 to 1 for any device

+++ android/src/main/java/org/reactnative/camera/CameraViewManager.java
@@ -86,7 +86,7 @@ public class CameraViewManager extends ViewGroupManager<RNCameraView> {
   }
 
   @ReactProp(name = "exposure")
-  public void setExposureCompensation(RNCameraView view, int exposure){
+  public void setExposureCompensation(RNCameraView view, float exposure){
     view.setExposureCompensation(exposure);
   }
